erstatt constate i HentOgSettBehandlingContext
